backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
httpx>=0.28.1
anthropic>=0.43.0
pydantic>=2.10.4
chromadb>=1.0.0
pypdf>=5.1.0
yfinance>=0.2.51
pandas>=2.2.3
sse-starlette>=2.2.1
beautifulsoup4>=4.12.3
lxml>=5.3.0
rapidfuzz>=3.0.0
pytest>=8.3.4
pytest-asyncio>=0.25.0
